split apm and deployment in deploymentresource to read and payload files
