+++ ec/ecresource/deploymentresource/apm/v2/apm_payload.go
@@ -23,9 +23,6 @@ import (
 	"github.com/elastic/cloud-sdk-go/pkg/models"
 	v1 "github.com/elastic/terraform-provider-ec/ec/ecresource/deploymentresource/apm/v1"
 	topologyv1 "github.com/elastic/terraform-provider-ec/ec/ecresource/deploymentresource/topology/v1"
-	"github.com/elastic/terraform-provider-ec/ec/ecresource/deploymentresource/utils"
-	"github.com/elastic/terraform-provider-ec/ec/internal/converters"
-	"github.com/elastic/terraform-provider-ec/ec/internal/util"
 	"github.com/hashicorp/terraform-plugin-framework/diag"
 	"github.com/hashicorp/terraform-plugin-framework/tfsdk"
 	"github.com/hashicorp/terraform-plugin-framework/types"
@@ -43,73 +40,6 @@ type ApmTF struct {
 	SizeResource              types.String `tfsdk:"size_resource"`
 	ZoneCount                 types.Int64  `tfsdk:"zone_count"`
 	Config                    types.Object `tfsdk:"config"`
-}
-
-type Apm struct {
-	ElasticsearchClusterRefId *string    `tfsdk:"elasticsearch_cluster_ref_id"`
-	RefId                     *string    `tfsdk:"ref_id"`
-	ResourceId                *string    `tfsdk:"resource_id"`
-	Region                    *string    `tfsdk:"region"`
-	HttpEndpoint              *string    `tfsdk:"http_endpoint"`
-	HttpsEndpoint             *string    `tfsdk:"https_endpoint"`
-	InstanceConfigurationId   *string    `tfsdk:"instance_configuration_id"`
-	Size                      *string    `tfsdk:"size"`
-	SizeResource              *string    `tfsdk:"size_resource"`
-	ZoneCount                 int        `tfsdk:"zone_count"`
-	Config                    *ApmConfig `tfsdk:"config"`
-}
-
-func ReadApms(in []*models.ApmResourceInfo) (*Apm, error) {
-	for _, model := range in {
-		if util.IsCurrentApmPlanEmpty(model) || utils.IsApmResourceStopped(model) {
-			continue
-		}
-
-		apm, err := ReadApm(model)
-		if err != nil {
-			return nil, err
-		}
-
-		return apm, nil
-	}
-
-	return nil, nil
-}
-
-func ReadApm(in *models.ApmResourceInfo) (*Apm, error) {
-	var apm Apm
-
-	apm.RefId = in.RefID
-	apm.ResourceId = in.Info.ID
-	apm.Region = in.Region
-	plan := in.Info.PlanInfo.Current.Plan
-
-	topologies, err := ReadApmTopologies(plan.ClusterTopology)
-	if err != nil {
-		return nil, err
-	}
-
-	if len(topologies) > 0 {
-		apm.InstanceConfigurationId = topologies[0].InstanceConfigurationId
-		apm.Size = topologies[0].Size
-		apm.SizeResource = topologies[0].SizeResource
-		apm.ZoneCount = topologies[0].ZoneCount
-	}
-
-	apm.ElasticsearchClusterRefId = in.ElasticsearchClusterRefID
-
-	apm.HttpEndpoint, apm.HttpsEndpoint = converters.ExtractEndpoints(in.Info.Metadata)
-
-	configs, err := readApmConfigs(plan.Apm)
-	if err != nil {
-		return nil, err
-	}
-
-	if len(configs) > 0 {
-		apm.Config = &configs[0]
-	}
-
-	return &apm, nil
 }
 
 func (apm ApmTF) Payload(ctx context.Context, payload models.ApmPayload) (*models.ApmPayload, diag.Diagnostics) {
@@ -170,7 +100,7 @@ func ApmPayload(ctx context.Context, apmObj types.Object, template *models.Deplo
 		return nil, nil
 	}
 
-	templatePayload := ApmResource(template)
+	templatePayload := payloadFromTemplate(template)
 
 	if templatePayload == nil {
 		diags.AddError("apm payload error", "apm specified but deployment template is not configured for it. Use a different template if you wish to add apm")
@@ -186,9 +116,9 @@ func ApmPayload(ctx context.Context, apmObj types.Object, template *models.Deplo
 	return payload, nil
 }
 
-// ApmResource returns the ApmPayload from a deployment
+// payloadFromTemplate returns the ApmPayload from a deployment
 // template or an empty version of the payload.
-func ApmResource(template *models.DeploymentTemplateInfoV2) *models.ApmPayload {
+func payloadFromTemplate(template *models.DeploymentTemplateInfoV2) *models.ApmPayload {
 	if template == nil || len(template.DeploymentTemplate.Resources.Apm) == 0 {
 		return nil
 	}
