+++ ec/ecresource/deploymentresource/read.go
@@ -143,7 +143,7 @@ func (r *Resource) read(ctx context.Context, id string, state *deploymentv2.Depl
 
 	deployment.ProcessSelfInObservability()
 
-	deployment.NullifyNotUsedEsTopologies(ctx, elasticsearchPlan)
+	deployment.NullifyUnusedEsTopologies(ctx, elasticsearchPlan)
 
 	// ReadDeployment returns empty config struct if there is no config, so we have to nullify it if plan doesn't contain it
 	// we use state for plan in Read and there is no state during import so we need to check elasticsearchPlan against nil
